+++ variables.tf
@@ -53,26 +53,26 @@ variable "subnet_ids" {
 
 // http://docs.aws.amazon.com/AmazonECS/latest/developerguide/ecs-optimized_AMI_launch_latest.html
 variable "ecs_optimized_amis" {
-  description = "List with ecs optimized images per region, last updated on: 2018-08-31 (2018.03.e)."
+  description = "List with ecs optimized images per region, last updated on: 2018-12-17 (2018.03.i)."
   type        = "map"
 
   default = {
-    us-east-1      = "ami-00129b193dc81bc31" # US East (N. Virginia)
-    us-east-2      = "ami-028a9de0a7e353ed9" # US East (Ohio)
-    us-west-1      = "ami-0d438d09af26c9583" # US West (N. California)
-    us-west-2      = "ami-00d4f478"          # US West (Oregon)
-    ca-central-1   = "ami-192fa27d"          # Canada (Central)
-    eu-west-1      = "ami-0af844a965e5738db" # EU (Ireland)
-    eu-west-2      = "ami-a44db8c3"          # EU (London)
-    eu-west-3      = "ami-07da674f0655ef4e1" # EU (Paris)
-    eu-central-1   = "ami-0291ba887ba0d515f" # EU (Frankfurt)
-    ap-northeast-1 = "ami-0041c416aa23033a2" # Asia Pacific (Tokyo)
-    ap-northeast-2 = "ami-047d2a61f94f862dc" # Asia Pacific (Seoul)
-    ap-southeast-1 = "ami-091bf462afdb02c60" # Asia Pacific (Singapore)
-    ap-southeast-2 = "ami-0092e55c70015d8c3" # Asia Pacific (Sydney)
-    ap-south-1     = "ami-0c179ca015d301829" # Asia Pacific (Mumbai)
-    sa-east-1      = "ami-0018ff8ee48970ac3" # South America (São Paulo)
-    us-gov-west-1  = "ami-c6079ba7"          # GovCloud
+    us-east-1      = "ami-045f1b3f87ed83659" # US East (N. Virginia)
+    us-east-2      = "ami-0307f7ccf6ea35750" # US East (Ohio)
+    us-west-1      = "ami-0285183bbef6224bd" # US West (N. California)
+    us-west-2      = "ami-01b70aea4161476b7" # US West (Oregon)
+    ca-central-1   = "ami-0f552e0a86f08b660" # Canada (Central)
+    eu-west-1      = "ami-0627e141ce928067c" # EU (Ireland)
+    eu-west-2      = "ami-01bee3897bba49d78" # EU (London)
+    eu-west-3      = "ami-0f4738fbeb53e6c3a" # EU (Paris)
+    eu-central-1   = "ami-0eaa3baf6969912ba" # EU (Frankfurt)
+    ap-northeast-1 = "ami-05b296a384694dfa4" # Asia Pacific (Tokyo)
+    ap-northeast-2 = "ami-00294948a592fc052" # Asia Pacific (Seoul)
+    ap-southeast-1 = "ami-050865a806e0dae53" # Asia Pacific (Singapore)
+    ap-southeast-2 = "ami-02c73ee1100ce3e7a" # Asia Pacific (Sydney)
+    ap-south-1     = "ami-01ef9f6a829ae3956" # Asia Pacific (Mumbai)
+    sa-east-1      = "ami-084b1eee100c102ee" # South America (São Paulo)
+    us-gov-west-1  = "ami-1dafcb7c"          # GovCloud
   }
 }